requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
torch>=1.12.0
transformers>=4.26.0
pydantic>=1.10.0
bitsandbytes>=0.39.0
PyJWT>=2.10.0
accelerate>=0.2.0
python-dotenv>=1.0.0